Fanarona Server Test Program
This code acts as a dummy client to test the handshake written in the
server implementation of the Fanarona game. Because this is a separate
'client', this must be compiled in a separate project, that way this
will run as a separate task. Finally, make sure to start the Fanarona
game process before starting the SocketTest process. This provides
almost no functionality, but serves as a test harness. Fanarona game in
server mode locks up before the handshake (expecting a connection on
the socket) without such a test harness.